c code specifies num_threads as u32

+++ bindings/c/src/lib.rs
@@ -60,10 +60,7 @@ impl Deref for DASContext {
 /// To avoid memory leaks, one should ensure that the pointer is freed after use
 /// by calling `eth_kzg_das_context_free`.
 #[no_mangle]
-pub extern "C" fn eth_kzg_das_context_new(
-    use_precomp: bool,
-    num_threads: usize,
-) -> *mut DASContext {
+pub extern "C" fn eth_kzg_das_context_new(use_precomp: bool, num_threads: u32) -> *mut DASContext {
     let use_precomp = if use_precomp {
         rust_eth_kzg::UsePrecomp::Yes {
             width: RECOMMENDED_PRECOMP_WIDTH,
@@ -75,7 +72,7 @@ pub extern "C" fn eth_kzg_das_context_new(
     let ctx = Box::new(DASContext {
         inner: rust_eth_kzg::DASContext::with_threads(
             &rust_eth_kzg::TrustedSetup::default(),
-            num_threads,
+            num_threads as usize,
             use_precomp,
         ),
     });
